requirements.txt
python-telegram-bot==20.7
openai==1.12.0
pytest==8.0.0
python-dotenv==1.0.1
boto3==1.34.34
requests==2.31.0
pytest-mock==3.12.0
responses==0.24.0 